task_modules/requirements.txt
# Environment variables
python-dotenv>=0.21.0

# AI API client
abacusai>=0.9.0

# Database
pymongo>=4.1.1

# Data processing
pandas>=1.5.0
numpy>=1.23.0

# For testing
pytest>=7.0.0
